+++ tests/testthat/test.NLCD.R
@@ -1,60 +1,61 @@
 library(FedData)
 library(httr)
+library(magrittr)
 context("National Land Cover Dataset tests")
 
-# test_that("The NLCD web coverage service is available at the correct URL", {
-#   skip_on_cran()
-#
-#   year <- 2016
-#   dataset <- "Land_Cover"
-#   landmass <- "L48"
-#
-#   coverage <- paste0("NLCD_", year, "_", dataset, "_", landmass)
-#   source <- "https://www.mrlc.gov/geoserver/wcs"
-#
-#   source %>%
-#     httr::modify_url(
-#       query = list(
-#         service = "WCS",
-#         version = "2.0.1",
-#         request = "DescribeCoverage",
-#         coverageid = coverage
-#       )
-#     )
-#
-#   cat(
-#     source %>%
-#       httr::modify_url(
-#         query = list(
-#           service = "WCS",
-#           version = "2.0.1",
-#           request = "DescribeCoverage",
-#           coverageid = coverage
-#         )
-#       ) %>%
-#       httr::GET() %>%
-#       httr::status_code(),
-#     "\n"
-#   )
-#
-#   expect_true(
-#     source %>%
-#       httr::modify_url(
-#         query = list(
-#           service = "WCS",
-#           version = "2.0.1",
-#           request = "DescribeCoverage",
-#           coverageid = coverage
-#         )
-#       ) %>%
-#       httr::GET() %>%
-#       httr::status_code() %>%
-#       identical(200L)
-#   )
-# })
+test_that("The NLCD web coverage service is available at the correct URL", {
+  skip_on_cran()
+
+  year <- 2016
+  dataset <- "Land_Cover"
+  landmass <- "L48"
+
+  coverage <- paste0("NLCD_", year, "_", dataset, "_", landmass)
+  source <- paste0("https://www.mrlc.gov/geoserver/mrlc_download/", coverage, "/wcs")
+
+  source %>%
+    httr::modify_url(
+      query = list(
+        service = "WCS",
+        version = "2.0.1",
+        request = "DescribeCoverage",
+        coverageid = coverage
+      )
+    )
+
+  cat(
+    source %>%
+      httr::modify_url(
+        query = list(
+          service = "WCS",
+          version = "2.0.1",
+          request = "DescribeCoverage",
+          coverageid = coverage
+        )
+      ) %>%
+      httr::GET() %>%
+      httr::status_code(),
+    "\n"
+  )
+
+  expect_true(
+    source %>%
+      httr::modify_url(
+        query = list(
+          service = "WCS",
+          version = "2.0.1",
+          request = "DescribeCoverage",
+          coverageid = coverage
+        )
+      ) %>%
+      httr::GET() %>%
+      httr::status_code() %>%
+      identical(200L)
+  )
+})
 
 test_that(
-  "The NLCD COG provides the same data as a raw download",
+  "The NLCD provides the same data as a raw download",
   {
     skip_on_cran()
 
@@ -70,14 +71,15 @@ test_that(
         template = raw,
         label = "PR",
         year = 2001,
-        landmass = "PR"
+        landmass = "PR",
+        force.redo = TRUE
       )
 
     raw %<>%
       raster::raster()
 
     expect_true(
-      all((feddata - raw)[] == 0)
+      all((feddata - raw)[] == 0, na.rm = TRUE)
     )
   }
 )